requirements.txt
Django==2.2.3
django-debug-toolbar==2.0
django-haystack==2.8.1
django-reversion==3.0.4
docutils
feincms3[all]==0.31.0
feincms3-meta==3.0.0
Whoosh==2.7.4
